quant_orchestrator/requirements.txt
temporalio
google-genai
pydantic
python-dotenv
pywinauto
